==> bonemarow_asapseq/code/42_makeResultingTabFig.R <==
library(BuenColors)

mdf <- readRDS("../output/Seurat_Proteinprojections_allMeta.rds")

kappa2(mdf[, c("archr_cluster", "predictions_l2")], weight = "unweighted")

dd <- mdf %>% group_by(predictions_l2, archr_cluster) %>%
  summarize(count = n()) %>% ungroup() %>% group_by(predictions_l2) %>% mutate(prop = count/sum(count)*100)

oo <- dd %>% arrange(desc(prop)) %>% pull(predictions_l2) %>% unique()
dd$order_l2 <- factor(as.character(dd$predictions_l2), levels = oo)

p1 <- ggplot(dd, aes(y = archr_cluster, x = order_l2, fill =  prop)) +
  geom_tile() + pretty_plot(fontsize = 6) + L_border() +
  scale_fill_gradientn(colours = jdb_palette("brewer_heat")) +
  theme(legend.position = "bottom") +
  scale_y_discrete(limits = rev(unique(dd$archr_cluster)))+ 
  labs(y = "ArchR Cluster (accessible chromatin only)", x = "Seurat Projection (Protein tags only)", fill= "% of Seurat projection") +
  theme(axis.text.x = element_text(angle = 45,  hjust = 1))

cowplot::ggsave2(p1, file = "../plots/heatmap_proportion_supplement.pdf",
                 width = 4, height = 2.7)


d2 <- mdf %>% group_by(predictions_l2, archr_cluster) %>%
  summarize(count = n()) %>% ungroup() %>% group_by(archr_cluster) %>% mutate(prop = count/sum(count)*100)

p2 <- ggplot(d2, aes(x = archr_cluster, y = predictions_l2, fill =  prop)) +
  geom_tile() + pretty_plot() + L_border() +
  scale_fill_gradientn(colours = jdb_palette("brewer_heat")) +
  theme(legend.position = "bottom") +
  scale_y_discrete(limits = rev(unique(d2$predictions_l2)))+ 
  labs(x = "ArchR Cluster", y = "Seurat Projection", fill= "% of ArchR cluster")

cowplot::ggsave2(cowplot::plot_grid(p1, p2, nrow =1), file = "../plots/heatmaps_proportions.png",
                 width = 12, height = 5)
